[ISSUE_70] - repackage to use SET proper packate name
