--- src/main/java/org/jboss/pull/processor/ProcessorException.java
@@ -1,12 +1,0 @@
-package org.jboss.pull.processor;
-
-
-public class ProcessorException extends Exception {
-
-    private static final long serialVersionUID = 1L;
-
-    public ProcessorException(String message, Throwable ex) {
-        super(message, ex);
-    }
-
-}
