--- src/main/java/org/jboss/pull/processor/data/Attribute.java
@@ -1,16 +1,0 @@
-package org.jboss.pull.processor.data;
-
-
-public class Attribute<T> {
-
-    private String name;
-
-    public Attribute(String name) {
-        this.name = name;
-    }
-
-    public String name() {
-        return name;
-    }
-
-}
