+++ src/main/resources/META-INF/services/org.jboss.pull.processor.Action
@@ -1,2 +1,2 @@
-org.jboss.pull.processor.impl.action.ReportAction
-org.jboss.pull.processor.impl.action.SetLabelsAction
+org.jboss.set.pull.processor.impl.action.ReportAction
+org.jboss.set.pull.processor.impl.action.SetLabelsAction
