+++ src/main/resources/META-INF/services/org.jboss.pull.processor.Evaluator
@@ -1,9 +1,9 @@
-org.jboss.pull.processor.impl.evaluator.BranchEvaluator
-org.jboss.pull.processor.impl.evaluator.IssuesRelatedEvaluator
-org.jboss.pull.processor.impl.evaluator.LabelsEvaluator
-org.jboss.pull.processor.impl.evaluator.PullRequestEvaluator
-org.jboss.pull.processor.impl.evaluator.PullRequestRelatedEvaluator
-org.jboss.pull.processor.impl.evaluator.RepositoryEvaluator
-org.jboss.pull.processor.impl.evaluator.StreamsEvaluator
-org.jboss.pull.processor.impl.evaluator.UpstreamEvaluator
-org.jboss.pull.processor.impl.evaluator.WritableRepositoryEvaluator
+org.jboss.set.pull.processor.impl.evaluator.BranchEvaluator
+org.jboss.set.pull.processor.impl.evaluator.IssuesRelatedEvaluator
+org.jboss.set.pull.processor.impl.evaluator.LabelsEvaluator
+org.jboss.set.pull.processor.impl.evaluator.PullRequestEvaluator
+org.jboss.set.pull.processor.impl.evaluator.PullRequestRelatedEvaluator
+org.jboss.set.pull.processor.impl.evaluator.RepositoryEvaluator
+org.jboss.set.pull.processor.impl.evaluator.StreamsEvaluator
+org.jboss.set.pull.processor.impl.evaluator.UpstreamEvaluator
+org.jboss.set.pull.processor.impl.evaluator.WritableRepositoryEvaluator
