+++ src/main/resources/META-INF/services/org.jboss.pull.processor.Processor
@@ -1,1 +1,1 @@
-org.jboss.pull.processor.impl.process.SETProcessor
+org.jboss.set.pull.processor.impl.process.SETProcessor
